requirements.txt
asgiref==3.5.2
Brotli==1.0.9
cffi==1.15.1
cssselect2==0.7.0
Django==4.1.2
django-ckeditor==6.5.1
django-cleanup==7.0.0
django-js-asset==2.0.0
django-widget-tweaks==1.4.12
fonttools==4.38.0
html5lib==1.1
Pillow==9.4.0
pycparser==2.21
pydyf==0.5.0
pyphen==0.13.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tinycss2==1.2.1
tzdata==2022.5
weasyprint==58.0
webencodings==0.5.1
zopfli==0.2.2
